moved files around and updated paths for better readability
