+++ api/database.json
@@ -23,24 +23,131 @@
   ],
   "articles": [
     {
-      "userId": 2,
-      "title": "",
-      "summary": "haha",
-      "date": "08/15/2018 at 13:14:24 PM",
-      "url": "",
+      "userId": 1,
+      "title": "Hey Alejandro v2.3.5.6.65.9",
+      "summary": "heyaksjdnflkajsdhflkjashfn v2.3.5.6.7.9.10",
+      "date": "edited on: 08/14/2018 at 15:20:56 PM",
+      "url": "jerk.com",
       "id": 1
+    },
+    {
+      "userId": 1,
+      "title": "sa,djbf,asd",
+      "summary": "dsbfmsad,bfasdf",
+      "date": "08/14/2018 at 15:29:33 PM",
+      "url": "sdfb,asdbf,s",
+      "id": 2
+    },
+    {
+      "userId": 2,
+      "title": "im super stupid",
+      "summary": "yes i am",
+      "date": "08/14/2018 at 15:36:26 PM",
+      "url": "asd,jbasdf.com",
+      "id": 3
+    },
+    {
+      "userId": 1,
+      "title": "Sup man",
+      "summary": "sadljkhfaksdhf;lsdf",
+      "date": "08/15/2018 at 9:37:53 AM",
+      "url": "asdlkfhakldsbfh.com",
+      "id": 4
+    },
+    {
+      "userId": 1,
+      "title": "test test test",
+      "summary": "adkfhlkqgflqwef",
+      "date": "08/15/2018 at 9:38:52 AM",
+      "url": "test.com",
+      "id": 5
+    },
+    {
+      "userId": 1,
+      "title": "dsFSDFFASDF",
+      "summary": "hgjlerbvhbepcqbr",
+      "date": "08/15/2018 at 9:40:18 AM",
+      "url": "asdljbadsb.com",
+      "id": 6
+    },
+    {
+      "userId": 1,
+      "title": "asdkfhldshf",
+      "summary": "a,sdhbalsdjbvlsdf",
+      "date": "08/15/2018 at 9:40:30 AM",
+      "url": "sadkjbfds.bc/sDF",
+      "id": 7
+    },
+    {
+      "userId": 1,
+      "title": "UAHRG,JAHLUGHKD.GF",
+      "summary": "ASDMFBASJKDGBVA,SGF",
+      "date": "08/15/2018 at 9:40:49 AM",
+      "url": "ASDJGDGF,AJSBCBJA.R.COM",
+      "id": 8
+    },
+    {
+      "userId": 1,
+      "title": "ergwergtewg",
+      "summary": "regfwerggwerg",
+      "date": "08/15/2018 at 15:14:42 PM",
+      "url": "argwergergerg.com",
+      "id": 9
     }
   ],
   "events": [
     {
-      "userId": 2,
-      "title": "jkyfyj",
-      "location": "drtsdjdykk",
-      "date": "2018-08-17",
-      "id": 3
+      "userId": 1,
+      "title": "Going on a Date",
+      "location": "Fido",
+      "date": "2018-08-15",
+      "id": 1
     }
   ],
-  "tasks": [],
+  "tasks": [
+    {
+      "title": "Slap Alex v3.7",
+      "completeDate": "2018-08-16",
+      "completion": false,
+      "userId": 1,
+      "id": 1
+    },
+    {
+      "title": "Slap Alex Again",
+      "completeDate": "2018-08-16",
+      "completion": false,
+      "userId": 1,
+      "id": 2
+    },
+    {
+      "title": "Slap Alex Again again",
+      "completeDate": "2018-08-16",
+      "completion": false,
+      "userId": 1,
+      "id": 3
+    },
+    {
+      "title": "Slap Alex Again forever",
+      "completeDate": "2018-08-16",
+      "completion": true,
+      "userId": 1,
+      "id": 4
+    },
+    {
+      "title": "sgfgsg",
+      "completeDate": "2018-08-16",
+      "completion": true,
+      "userId": 1,
+      "id": 5
+    },
+    {
+      "title": "Slap the mess out of alex",
+      "completeDate": "2018-08-17",
+      "completion": false,
+      "userId": 1,
+      "id": 6
+    }
+  ],
   "firends": [],
   "messages": [
     {
