+++ src/css/main.css
@@ -68,6 +68,24 @@ main {
     color: blue;
 }
 
+#profile-content {
+    background-color: white;
+    margin: .3em;
+    padding: .5em;
+}
+
+#event-content {
+    background-color: lightgray;
+    margin: .3em;
+    padding: .5em;
+}
+
+#task-main {
+    background-color: lightseagreen;
+    margin: .3em;
+    padding: .5em;
+}
+
 #profile-image {
     display: block;
     width: 100%;
