+++ src/js/articles/articleForm.js
@@ -20,19 +20,5 @@ const articleFormManager = Object.create(null, {
     }
 });
 
-// function renderFormBtn(){
-//     document.querySelector("#article-form-container").innerHTML +=
-//         `<button id="add-article-btn">Add A New Article</button>`
-// }
-
-// function renderArticleForm(){
-//     document.querySelector("#article-form").innerHTML =
-//     `<div class="article-form">
-//         <input placeholder="Article Title">
-//         <textarea placeholder="Summary of Article"></textarea>
-//         <input placeholder="Paste a link to your article here">
-//         <button id="post-article">Post Article</button>
-//     </div>`
-// }
 
 module.exports = articleFormManager;
